+++ packages/app/src/components/ShareLink/ShareLink.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next';
 import { toastSuccess, toastError } from '~/client/util/apiNotification';
 import { apiv3Delete } from '~/client/util/apiv3-client';
 import { useCurrentPageId } from '~/stores/context';
-import { useSWRxSharelink } from '~/stores/sharelink';
+import { useSWRxSharelink } from '~/stores/share-link';
 
 import ShareLinkForm from './ShareLinkForm';
 import ShareLinkList from './ShareLinkList';
